+++ src/core/QueryManager.ts
@@ -86,6 +86,7 @@ import {
 import { tryFunctionOrLogError } from '../util/errorHandling';
 
 import {
+  isApolloError,
   ApolloError,
 } from '../errors/ApolloError';
 
@@ -538,6 +539,7 @@ export class QueryManager {
         document: queryDoc,
         complete: !shouldFetch,
         queryId,
+        requestId,
       });
     }
 
@@ -994,7 +996,7 @@ export class QueryManager {
         }).catch((error: Error) => {
           // This is for the benefit of `refetch` promises, which currently don't get their errors
           // through the store like watchQuery observers do
-          if (error instanceof ApolloError) {
+          if (isApolloError(error)) {
             reject(error);
           } else {
             this.store.dispatch({